[refactor] #50 불 필요한 auth 의존성 제거
